+++ src/commands/builders/getRoles.ts
@@ -1,5 +1,5 @@
 import { PermissionsBitField, SlashCommandBuilder } from 'discord.js';
-import lang from "../../utils/lang.json";
+import lang from '../../utils/lang.json';
 
 /* main slash command */
 export const getRoles = new SlashCommandBuilder()