+++ src/commands/handlers/removeRole.ts
@@ -1,7 +1,7 @@
-import { Client, ChatInputCommandInteraction, CacheType } from "discord.js";
-import { AppDataSource } from "../../typeorm";
-import { SavedRole } from "../../typeorm/entities/SavedRole";
-import lang from "../../utils/lang.json";
+import { CacheType, ChatInputCommandInteraction, Client } from 'discord.js';
+import { AppDataSource } from '../../typeorm';
+import { SavedRole } from '../../typeorm/entities/SavedRole';
+import lang from '../../utils/lang.json';
 
 const savedRoleRepo = AppDataSource.getRepository(SavedRole);
 
@@ -44,8 +44,8 @@ export const removeRoleHandler = async(client: Client, interaction: ChatInputCom
 
             savedRoleRepo.createQueryBuilder()
                 .delete()
-                .where("role = :role", { role: role })
-                .andWhere("type = :type", { type: 'staff' })
+                .where('role = :role', { role: role })
+                .andWhere('type = :type', { type: 'staff' })
                 .execute();
 
             // after completion, send an ephemeral success message
@@ -67,8 +67,8 @@ export const removeRoleHandler = async(client: Client, interaction: ChatInputCom
 
             savedRoleRepo.createQueryBuilder()
                 .delete()
-                .where("role = :role", { role: role })
-                .andWhere("type = :type", { type: 'admin' })
+                .where('role = :role', { role: role })
+                .andWhere('type = :type', { type: 'admin' })
                 .execute();
             
             // after completion, send an ephemeral success message
@@ -86,4 +86,4 @@ export const removeRoleHandler = async(client: Client, interaction: ChatInputCom
             ephemeral: true
         });
     }
-}
+};