+++ src/commands/handlers/ticketReply.ts
@@ -1,7 +1,7 @@
-import { Client, ChatInputCommandInteraction, CacheType, TextChannel } from 'discord.js';
+import { CacheType, ChatInputCommandInteraction, Client, TextChannel } from 'discord.js';
 import { AppDataSource } from '../../typeorm';
-import { Ticket } from "../../typeorm/entities/Ticket";
-import lang from "../../utils/lang.json";
+import { Ticket } from '../../typeorm/entities/Ticket';
+import lang from '../../utils/lang.json';
 
 const ticketRepo = AppDataSource.getRepository(Ticket);
 
@@ -27,14 +27,14 @@ export const ticketReplyHandler = async(client: Client, interaction: ChatInputCo
         // replies for ban appeals
         case 'bapple': {
             if (subcommand == 'approve') { 
-                await channel.send(`<@${ticketUser}>` + ", " + lang.ticketReply.bapple.approve);
+                await channel.send(`<@${ticketUser}>` + ', ' + lang.ticketReply.bapple.approve);
                 await interaction.reply({
                     content: lang.ticketReply.bapple.approveSent,
                     ephemeral: true
                 });
             } 
             else if (subcommand == 'deny') { 
-                await channel.send(`<@${ticketUser}>` + ", " + lang.ticketReply.bapple.deny);
+                await channel.send(`<@${ticketUser}>` + ', ' + lang.ticketReply.bapple.deny);
                 await interaction.reply({
                     content: lang.ticketReply.bapple.denySent,
                     ephemeral: true
@@ -42,4 +42,4 @@ export const ticketReplyHandler = async(client: Client, interaction: ChatInputCo
             }
         }
     }
-}
+};