+++ src/events/ticket/ageVerify.ts
@@ -1,4 +1,4 @@
-import { ActionRowBuilder, ModalBuilder, TextInputStyle, TextInputBuilder, ModalSubmitFields} from "discord.js";
+import { ActionRowBuilder, ModalBuilder, ModalSubmitFields, TextInputBuilder, TextInputStyle } from 'discord.js';
 
 export const ageVerifyModal = async(modal: ModalBuilder) => {
     const verifyModal = new ActionRowBuilder<TextInputBuilder>().addComponents(
@@ -10,10 +10,10 @@ export const ageVerifyModal = async(modal: ModalBuilder) => {
     );
 
     return modal.addComponents(verifyModal);
-}
+};
 
 export const ageVerifyMessage = async(fields: ModalSubmitFields) => {
-    const header = `# 18+ Verify\n`;
+    const header = '# 18+ Verify\n';
     const vdob = `Date of Birth: ${fields.getTextInputValue('dob_input')}`;
     return header + vdob;
-}
+};