+++ src/events/ticket/bugReport.ts
@@ -1,4 +1,4 @@
-import { ActionRowBuilder, ModalBuilder, TextInputStyle, TextInputBuilder, ModalSubmitFields} from "discord.js";
+import { ActionRowBuilder, ModalBuilder, ModalSubmitFields, TextInputBuilder, TextInputStyle } from 'discord.js';
 
 export const bugReportModal = async(modal: ModalBuilder) => {
     const brModalI = new ActionRowBuilder<TextInputBuilder>().addComponents(
@@ -10,10 +10,10 @@ export const bugReportModal = async(modal: ModalBuilder) => {
     );
 
     modal.addComponents(brModalI);
-}
+};
 
 export const bugReportMessage = async(fields: ModalSubmitFields) => {
-    const header = `# Bug Report\n`;
-    const brd = `Report Description: ${fields.getTextInputValue('bug_report_input')}\n`
+    const header = '# Bug Report\n';
+    const brd = `Report Description: ${fields.getTextInputValue('bug_report_input')}\n`;
     return header + brd;
-}
+};