+++ src/typeorm/entities/ArchivedTicket.ts
@@ -1,4 +1,4 @@
-import { Column, Entity, PrimaryGeneratedColumn } from "typeorm";
+import { Column, Entity, PrimaryGeneratedColumn } from 'typeorm';
 
 @Entity({ name: 'archived_tickets'})
 export class ArchivedTicket {